pyproject.toml
[project]
name = "outpost"
version = "0.1.0"
description = "Spin up and spin down DigitalOcean droplets as remote outposts."
authors = [{ name = "kr-b" }]
dependencies = [
    "click",
    "pydo",
    "python-dotenv"
]


[project.scripts]
outpost = "outpost.cli:cli"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"